agent/pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neos-agent"
version = "0.1.0"
description = "NEOS governance agent webservice"
requires-python = ">=3.12"
dependencies = [
    "sanic>=25.0,<26.0",
    "sanic-ext>=25.0,<26.0",
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic-settings>=2.0",
    "anthropic>=0.40",
    "datastar-py>=0.8",
    "jinja2>=3.1",
    "PyNaCl>=1.5",
    "base58>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "sanic-testing>=24.6",
    "aiosqlite>=0.20",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
